Plugins: categorize windows.drivermodule as a malware plugin
